backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
passlib
python-jose
alembic
python-multipart
python-dotenv
google-generativeai
fastapi-session==0.2.7
google-cloud-speech
google-cloud-texttospeech
httpx
authlib
itsdangerous
